requirements.txt
streamlit==1.27.0
openai==1.3.0
python-dotenv==1.0.0
pillow==11.1.0